+++ docutils/docutils/parsers/rst/directives/misc.py
@@ -14,7 +14,7 @@ from pathlib import Path
 from urllib.request import urlopen
 from urllib.error import URLError
 
-from docutils import io, nodes, statemachine, utils
+from docutils import frontend, io, nodes, statemachine, utils
 from docutils.parsers.rst import Directive, convert_directive_function
 from docutils.parsers.rst import directives, roles, states
 from docutils.parsers.rst.directives.body import CodeBlock, NumberLines
@@ -213,14 +213,17 @@ class Include(Directive):
 
         Provisional.
         """
-        settings = self.settings.copy()
+        parser = self.options['parser']()
+        settings = frontend.get_default_settings(parser)
+        # update with current document settings
+        for k, v in self.settings.__dict__.items():
+            setattr(settings, k, v)
         settings._source = self.options['source']
         document = utils.new_document(settings._source, settings)
         document.include_log = self.state.document.include_log
         document.ids = self.state.document.ids
         document.nameids = self.state.document.nameids
         document.nametypes = self.state.document.nametypes
-        parser = self.options['parser']()
         parser.parse(text, document)
         self.state.document.parse_messages.extend(document.parse_messages)
         # clean up doctree and complete parsing